While running my_joystick as main pressed buttons are shown with their respective number, also the axes numbers are displayed. On top of the terminal also the name assigned to your joystick and your platform (corresponds to your OS) is displayed.
For new device you just have to create a new elif statement with joystick name and possibly platform name (I have the same joystick, recognized with the same name, but with different buttons assignment on Mac and Windows). Then you just need to complete the list of buttons according to what you see in terminal when pressing them.

All the messy stuff is shifted down and the same for all the joysticks.